+++ taboolib-ioc-annotation/src/main/kotlin/top/wcpe/taboolib/ioc/annotation/ComponentScan.kt
@@ -1,0 +1,11 @@
+package top.wcpe.taboolib.ioc.annotation
+
+import kotlin.reflect.KClass
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ComponentScan(
+    val value: Array<String> = [],
+    val basePackages: Array<String> = [],
+    val basePackageClasses: Array<KClass<*>> = []
+)