+++ taboolib-ioc-annotation/src/main/kotlin/top/wcpe/taboolib/ioc/annotation/Prototype.kt
@@ -1,0 +1,5 @@
+package top.wcpe.taboolib.ioc.annotation
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Prototype